BE/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pydantic-settings
python-dotenv
pandas
playwright
beautifulsoup4
requests
lxml
aiofiles
aiohttp
python-multipart
Jinja2
loguru
google.generativeai
asyncio